Launcher3: Define blur radius in device-independent pixels
Specifying the blur radius in pixels leads to inconsistent results
across devices with different pixel densities. With a Pixel 4 XL
(1440x3040) as the reference display (3.5x scale factor), 175 px is more
or less the amount of blur we want, but it's too strong on the Pixel 5
(1080x2340).

Define the blur radius in device-independent pixels (calculated from the
reference display described above) to make the strength uniform.

Signed-off-by: Adithya R <gh0strider.2k18.reborn@gmail.com>